pyproject.toml
[tool.poetry]
name = "desing_patterns"
version = "0.1.0"
description = ""
authors = ["Gonzalo Donoso <gdonoso94@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-interface = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
